% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{check_table_in_db}
\alias{check_table_in_db}
\title{table_in_db}
\usage{
check_table_in_db(db_file, schema, table)
}
\arguments{
\item{db_file}{Path to database file.}

\item{schema}{Schema name to check}

\item{table}{Table name to check}
}
\description{
Check if table exists in database.
}
