% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_parent_dir}
\alias{create_parent_dir}
\title{create_parent_dir}
\usage{
create_parent_dir(path)
}
\arguments{
\item{path}{Path to create, will only create if it does not exist}
}
\description{
create_parent_dir
}
