% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{get_connection}
\alias{get_connection}
\title{get_connection
Get connection to database.}
\usage{
get_connection(db, read_only = TRUE)
}
\arguments{
\item{db}{Path to database}

\item{read_only}{Whether to open connection in read-only mode.}
}
\value{
Connection to database
}
\description{
get_connection
Get connection to database.
}
