% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_dense_layer}
\alias{get_dense_layer}
\title{get_dense_layer
Convert a layer in Seurat object to dense data.table, with rows being cells.}
\usage{
get_dense_layer(obj, assay = NULL, layer = NULL)
}
\arguments{
\item{obj}{Seurat object}

\item{assay}{Which assay to use (e.g. "RNA")}

\item{layer}{Which layer to use (e.g. "counts")}
}
\value{
\code{data.table} with dense layer information.
}
\description{
get_dense_layer
Convert a layer in Seurat object to dense data.table, with rows being cells.
}
