% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_embedding_names}
\alias{get_embedding_names}
\title{get_embedding_names}
\usage{
get_embedding_names(obj)
}
\arguments{
\item{obj}{Seurat object}
}
\value{
Vector of the embedding names.
}
\description{
Get names of dimensionality reduction embeddings in Seurat object.
}
