% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{get_schemas_in_db}
\alias{get_schemas_in_db}
\title{Get Schemas in Database}
\usage{
get_schemas_in_db(db_file)
}
\arguments{
\item{db_file}{Path to database file.}
}
\value{
A data frame containing the names of the schemas in the database.
}
\description{
This function retrieves the names of all schemas
in the connected database, excluding the default schemas.
}
