% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{get_tables_in_schema}
\alias{get_tables_in_schema}
\title{Get Tables for a Specific Schema}
\usage{
get_tables_in_schema(db_file, schema)
}
\arguments{
\item{db_file}{Path to database file.}

\item{schema}{Schema to find tables in}
}
\value{
Character vector with table names, if any.
}
\description{
This function retrieves the names of all tables
in the connected database, selecting only tables matching a schema.
}
