% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{get_tables_per_schema_in_db}
\alias{get_tables_per_schema_in_db}
\title{Get Tables per Schema in Database}
\usage{
get_tables_per_schema_in_db(db_file)
}
\arguments{
\item{db_file}{Path to database file.}
}
\value{
A data frame containing the names of the tables and schemas in the database.
}
\description{
This function retrieves the names of all tables
in the connected database, listing them per schema.
}
