% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_data_with_meta}
\alias{read_data_with_meta}
\title{read_data_with_meta
Read count data of Seurat object and merge it with metadata.}
\usage{
read_data_with_meta(
  datadir = ".",
  what = "layer",
  name = "counts",
  col_select = NULL
)
}
\arguments{
\item{datadir}{Path to parent directory of feather files.}

\item{what}{Whether to read count data (\code{layer}) or an \code{embedding}.}

\item{name}{Which layer of data to read (e.g. \code{counts}).}

\item{col_select}{Optional vector of columns to read of layer/embedding.}
}
\value{
data.frame of read data
}
\description{
read_data_with_meta
Read count data of Seurat object and merge it with metadata.
}
