% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_layer}
\alias{read_layer}
\title{read_layer}
\usage{
read_layer(db_file, layer = "counts", col_select = NULL)
}
\arguments{
\item{db_file}{Path to database_file}

\item{layer}{Which layer to read.}

\item{col_select}{Optional vector of columns to read.}
}
\value{
tibble of read data
}
\description{
Read count data of Seurat object.
}
