% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_metadata}
\alias{read_metadata}
\title{read_metadata}
\usage{
read_metadata(db_file)
}
\arguments{
\item{db_file}{Path to database file.}
}
\value{
tibble of read data
}
\description{
Read metadata of Seurat object.
}
