% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{return_tibble}
\alias{return_tibble}
\title{return_tibble
Collect and return data as tibble.}
\usage{
return_tibble(table, col_select = NULL)
}
\arguments{
\item{table}{Table to collect}

\item{col_select}{Optional vector of columns to select.}
}
\value{
tibble of collected data
}
\description{
return_tibble
Collect and return data as tibble.
}
