% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{subset_data}
\alias{subset_data}
\title{subset_data
Subset data based on values. Convenience wrapper.}
\usage{
subset_data(data, subset_col, subset_values)
}
\arguments{
\item{data}{The data.frame to subset}

\item{subset_col}{Which column to use for filtering.}

\item{subset_values}{Which values to keep.}
}
\description{
subset_data
Subset data based on values. Convenience wrapper.
}
