% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{write_seurat_embeddings}
\alias{write_seurat_embeddings}
\title{write_seurat_embeddings}
\usage{
write_seurat_embeddings(obj, con = ".", layers = NULL)
}
\arguments{
\item{obj}{Seurat object.}

\item{con}{Connection to database.}

\item{layers}{Which layers to export. Leave empty for all.}
}
\description{
Write out all data stored in embeddings, like PCA and UMAP coordinates
}
