% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{write_seurat_metadata}
\alias{write_seurat_metadata}
\title{write_seurat_metadata
Write out all data stored in metadata of Seurat object.}
\usage{
write_seurat_metadata(obj, outdir = ".")
}
\arguments{
\item{obj}{Seurat object.}

\item{outdir}{Path to output directory.}
}
\value{
Vector of written file paths.
}
\description{
write_seurat_metadata
Write out all data stored in metadata of Seurat object.
}
