% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{write_seurat_to_db}
\alias{write_seurat_to_db}
\title{write_seurat_to_db}
\usage{
write_seurat_to_db(obj, db = "seurat.duckdb")
}
\arguments{
\item{obj}{Seurat object.}

\item{db}{Path to database}
}
\value{
Vector of written file paths.
}
\description{
Write out all data stored in Seurat object.
}
