% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{write_seurat_to_feather}
\alias{write_seurat_to_feather}
\title{write_seurat_to_feather
Write out all data stored in Seurat object.}
\usage{
write_seurat_to_feather(obj, outdir = ".")
}
\arguments{
\item{obj}{Seurat object.}

\item{outdir}{Path to output directory.}
}
\value{
Vector of written file paths.
}
\description{
write_seurat_to_feather
Write out all data stored in Seurat object.
}
