% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{check_table_in_db}
\alias{check_table_in_db}
\title{table_in_db
Check if table exists in database.}
\usage{
check_table_in_db(con, schema, table)
}
\arguments{
\item{con}{Connection to database}

\item{schema}{Schema name to check}

\item{table}{Table name to check}
}
\description{
table_in_db
Check if table exists in database.
}
