% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{get_connection}
\alias{get_connection}
\title{get_connection}
\usage{
get_connection(db_file, read_only = TRUE)
}
\arguments{
\item{db_file}{Path to database}

\item{read_only}{Whether to open connection in read-only mode.}
}
\value{
Connection to database
}
\description{
Get connection to database.
}
