% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_data_with_meta}
\alias{read_data_with_meta}
\title{read_data_with_meta}
\usage{
read_data_with_meta(
  db_file,
  what = "layer",
  name = "counts",
  col_select = NULL
)
}
\arguments{
\item{what}{Whether to read count data (\code{layer}) or an \code{embedding}.}

\item{name}{Which layer of data to read (e.g. \code{counts}).}

\item{col_select}{Optional vector of columns to read of layer/embedding.}

\item{Path}{to database file.}
}
\value{
data.frame of read data
}
\description{
Read count data of Seurat object and merge it with metadata.
}
