% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_embedding}
\alias{read_embedding}
\title{read_embedding
Read embedding data of Seurat object.}
\usage{
read_embedding(con, layer = "umap", col_select = NULL)
}
\arguments{
\item{con}{Connection to database}

\item{layer}{Which layer to read.}

\item{col_select}{Optional vector of columns to read.}
}
\value{
tibble of read data
}
\description{
read_embedding
Read embedding data of Seurat object.
}
