% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_embedding}
\alias{read_embedding}
\title{read_embedding}
\usage{
read_embedding(db_file, layer = "umap", col_select = NULL)
}
\arguments{
\item{db_file}{Path to database file.}

\item{layer}{Which layer to read.}

\item{col_select}{Optional vector of columns to read.}
}
\value{
tibble of read data
}
\description{
Read embedding data of Seurat object.
}
