% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_layer}
\alias{read_layer}
\title{read_layer
Read count data of Seurat object.}
\usage{
read_layer(datadir = ".", layer = "counts", col_select = NULL)
}
\arguments{
\item{datadir}{Path to parent directory of feather files.}

\item{layer}{Which layer to read.}

\item{col_select}{Optional vector of columns to read.}
}
\value{
data.frame of read data
}
\description{
read_layer
Read count data of Seurat object.
}
