% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_metadata}
\alias{read_metadata}
\title{read_metadata}
\usage{
read_metadata(con)
}
\arguments{
\item{con}{Connection to database}
}
\value{
tibble of read data
}
\description{
Read metadata of Seurat object.
}
