% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_metadata}
\alias{read_metadata}
\title{read_metadata
Read metadata of Seurat object.}
\usage{
read_metadata(con)
}
\arguments{
\item{con}{Connection to database}

\item{col_select}{Optional vector of columns to read.}
}
\value{
tibble of read data
}
\description{
read_metadata
Read metadata of Seurat object.
}
