% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{which_duplicated_columns}
\alias{which_duplicated_columns}
\title{which_duplicated_columns}
\usage{
which_duplicated_columns(df)
}
\arguments{
\item{df}{The data.frame to check}
}
\value{
Vector with names of duplicated columns.
}
\description{
Check if there are duplicated columns in a data.frame.
}
