% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{write_seurat_counts}
\alias{write_seurat_counts}
\title{write_seurat_counts
Write out all data stored in layers, like counts and scaled data.}
\usage{
write_seurat_counts(obj, outdir = ".", layers = NULL)
}
\arguments{
\item{obj}{Seurat object.}

\item{outdir}{Path to output directory.}

\item{layers}{Which layers to export. Leave empty for all.}
}
\value{
Vector of written file paths.
}
\description{
write_seurat_counts
Write out all data stored in layers, like counts and scaled data.
}
