% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{write_seurat_embeddings}
\alias{write_seurat_embeddings}
\title{write_seurat_embeddings
Write out all data stored in embeddings, like PCA and UMAP coordinates}
\usage{
write_seurat_embeddings(obj, outdir = ".", layers = NULL)
}
\arguments{
\item{obj}{Seurat object.}

\item{outdir}{Path to output directory.}

\item{layers}{Which layers to export. Leave empty for all.}
}
\value{
Vector of written file paths.
}
\description{
write_seurat_embeddings
Write out all data stored in embeddings, like PCA and UMAP coordinates
}
