% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{write_seurat_metadata}
\alias{write_seurat_metadata}
\title{write_seurat_metadata}
\usage{
write_seurat_metadata(obj, con = ".")
}
\arguments{
\item{obj}{Seurat object.}

\item{con}{Connection to database.}
}
\description{
Write out all data stored in metadata of Seurat object.
}
